moving StagingSlotName to opentrons.types
